@GitHK review: refactor to slim gc plugin
