+++ services/web/server/src/simcore_service_webserver/resource_manager/config.py
@@ -9,13 +9,13 @@ from servicelib.aiohttp.application_keys import APP_CONFIG_KEY
 
 CONFIG_SECTION_NAME = "resource_manager"
 
-APP_CLIENT_REDIS_CLIENT_KEY = __name__ + ".resource_manager.redis_client"
-APP_CLIENT_REDIS_LOCK_MANAGER_KEY = __name__ + ".resource_manager.redis_lock"
+APP_CLIENT_REDIS_CLIENT_KEY = f"{__name__}.resource_manager.redis_client"
+APP_CLIENT_REDIS_LOCK_MANAGER_KEY = f"{__name__}.resource_manager.redis_lock"
 APP_CLIENT_REDIS_LOCK_MANAGER_CLIENT_KEY = (
-    __name__ + ".resource_manager.redis_lock_client"
+    f"{__name__}.resource_manager.redis_lock_client"
 )
-APP_CLIENT_SOCKET_REGISTRY_KEY = __name__ + ".resource_manager.registry"
-APP_RESOURCE_MANAGER_TASKS_KEY = __name__ + ".resource_manager.tasks.key"
+APP_CLIENT_SOCKET_REGISTRY_KEY = f"{__name__}.resource_manager.registry"
+APP_RESOURCE_MANAGER_TASKS_KEY = f"{__name__}.resource_manager.tasks.key"
 
 
 # lock names and format strings
